+++ Changelog.md
@@ -2,11 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.6.0] - 2025-11-15
+
+### Added
+
+- Flowers Block
+- Handled the event when the timer ends
+- Added stats view in settings
+
+---
+
 ## [0.5.1] - 2025-11-15
 
 ### Fixes
 
 - Text alignment in settings ui
+
+---
 
 ## [0.5.0] - 2025-11-15
 
@@ -21,6 +33,8 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Lowered hances of a grass block appearing if the user only does a short pomodoro session
+
+---
 
 ## [0.4.0] - 2025-11-15
 
@@ -38,6 +52,8 @@ All notable changes to this project will be documented in this file.
 
 - No longer passing full pomodoro session time, now its the duration the user studied for
 
+---
+
 ## [0.3.0] - 2025-11-15
 
 ### Added
@@ -53,6 +69,8 @@ All notable changes to this project will be documented in this file.
 - Inverted camera movement with cursor on y axis
 - Spelling Mistake in title in menu
 
+---
+
 ## [0.2.0] - 2025-11-15
 
 ### Added
@@ -61,12 +79,16 @@ All notable changes to this project will be documented in this file.
 - Upgraded timer input
 - After each pomodoro session a grass block is added
 
+---
+
 ## [0.1.0] - 2025-11-14
 
 ### Added
 
 - Changed menu camera position
 - Added grid system for blocks
+
+---
 
 ## [0.0.1] - 2025-11-14